fix(search): satisfy macOS semantic clippy
Amp-Thread-ID: https://ampcode.com/threads/T-019fff40-1504-7718-bcc2-52625d496ba3

+++ src-tauri/src/semantic_search.rs
@@ -433,7 +433,7 @@ impl SystemEmbeddingProvider {
                             .to_string(),
                     )
                 })?;
-            let dimension = unsafe { embedding.dimension() } as usize;
+            let dimension = unsafe { embedding.dimension() };
             let revision = unsafe { embedding.revision() };
             if dimension == 0 || dimension > MAX_EMBEDDING_DIMENSION {
                 return Err(ProviderFailure::Unavailable(
